todo: edit profile

+++ src/network/router.cpp
@@ -159,7 +159,7 @@ void Router::handlePacket(const QByteArray& rawPacket)
             // Add the uid of sender to jsonData
             QJsonArray arr = QJsonDocument::fromJson(jsonData.toUtf8()).array();
             arr.prepend(
-                (int)qobject_cast<ServerPlayer *>(parent())->getId()
+                qobject_cast<ServerPlayer *>(parent())->getId()
             );
             ServerInstance->callLua(command, QJsonDocument(arr).toJson());
         }
